+++ apps/meteor/client/views/room/body/RoomBodyV2.tsx
@@ -6,7 +6,7 @@ import { memo, useCallback, useMemo } from 'react';
 import { isTruthy } from '../../../../lib/isTruthy';
 import { CustomScrollbars } from '../../../components/CustomScrollbars';
 import { useEmbeddedLayout } from '../../../hooks/useEmbeddedLayout';
-import { useMergedRefsV2 } from '../../../hooks/useMessageComposerMergedRefs';
+import { useMergedRefsV2 } from '../../../hooks/useMergedRefsV2';
 import { BubbleDate } from '../BubbleDate';
 import { MessageList } from '../MessageList';
 import DropTargetOverlay from './DropTargetOverlay';
